blockscout_agent/requirements.txt
# google-genai
# google-adk
python-dotenv
openai-agents
openai 